+++ ManagedCode.Repository.MongoDB/IMongoDbRepository.cs
@@ -4,5 +4,6 @@ using MongoDB.Bson;
 namespace ManagedCode.Repository.MongoDB
 {
     public interface IMongoDbRepository<TItem> : IRepository<ObjectId, TItem> where TItem : class, IItem<ObjectId>
-    {}
+    {
+    }
 }
