+++ ManagedCode.Repository.Tests/Common/TestAzureTableItem.cs
@@ -1,8 +1,13 @@
+using System;
+using ManagedCode.Repository.AzureTable;
+
 namespace ManagedCode.Repository.Tests.Common
 {
-    public class TestAzureTableItem : AzureTable.AzureTableItem
+    public class TestAzureTableItem : AzureTableItem, IBaseItem<TableId>
     {
-        public string Data { get; set; }
+        public string StringData { get; set; }
         public int IntData { get; set; }
+        public float FloatData { get; set; }
+        public DateTime DateTimeData { get; set; }
     }
 }
