+++ ManagedCode.Repository.Tests/Common/TestMongoDbItem.cs
@@ -1,26 +1,14 @@
+using System;
+using ManagedCode.Repository.MongoDB;
 using MongoDB.Bson;
 
 namespace ManagedCode.Repository.Tests.Common
 {
-    public class TestMongoDbItem : ManagedCode.Repository.MongoDB.MongoDbItem
+    public class TestMongoDbItem : MongoDbItem, IBaseItem<ObjectId>
     {
-        public TestMongoDbItem() : base()
-        {
-        }
-
-        public TestMongoDbItem(string id) : base(id)
-        {
-        }
-
-        public TestMongoDbItem(string id, string partKey) : base(id)
-        {
-            Id = ObjectId.Parse(id);
-            PartKey = partKey;
-        }
-
-        public string PartKey { get; set; }
-        public string Data { get; set; }
-        public string RowKey { get; set; }
+        public string StringData { get; set; }
         public int IntData { get; set; }
+        public float FloatData { get; set; }
+        public DateTime DateTimeData { get; set; }
     }
 }
